+++ CHANGELOG.md
@@ -13,8 +13,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
  - Exposed `Enum` class
+  - Used `cross-spawn` module for Windows compatibility
  - Updated `sqlite3` to `4.1.1`
- - Used `cross-spawn` module for Windows compatibility
+ - Updated `inquirer` to `7.0.1`
+ - Updated `knex` to `0.20.4`
+ - Updated `nodemon` to `2.0.2`
+ - Updated `replace-in-file` to `5.0.2`
 
 ### Fixed
  - Nested translation files can be used with dot syntax