+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+
+## [1.0.0-rc.3] - 2019-12-19
+
 ### Changed
  - Exposed `Enum` class
   - Used `cross-spawn` module for Windows compatibility
@@ -597,7 +601,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
-[Unreleased]:    https://github.com/absolunet/node-ioc/compare/1.0.0-rc.2...HEAD
+[Unreleased]:    https://github.com/absolunet/node-ioc/compare/1.0.0-rc.3...HEAD
+[1.0.0-rc.3]:    https://github.com/absolunet/node-ioc/compare/1.0.0-rc.2...1.0.0-rc.3
 [1.0.0-rc.2]:    https://github.com/absolunet/node-ioc/compare/1.0.0-rc.1...1.0.0-rc.2
 [1.0.0-rc.1]:    https://github.com/absolunet/node-ioc/compare/1.0.0-beta.3...1.0.0-rc.1
 [1.0.0-beta.3]:  https://github.com/absolunet/node-ioc/compare/1.0.0-beta.2...1.0.0-beta.3